customer/src/main/java/nl/alimjan/customer/CustomerJPADaoService.java
package nl.alimjan.customer;

import java.util.List;
import java.util.Optional;
import org.springframework.stereotype.Repository;

@Repository("jpa")
public class CustomerJPADaoService implements CustomerDao {

    private final CustomerRepository customerRepository;

    public CustomerJPADaoService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    @Override
    public List<Customer> selectAllCustomers() {
        return customerRepository.findAll();
    }

    @Override
    public Optional<Customer> selectCustomerById(Long id) {
        return customerRepository.findById(id);
    }

    @Override
    public Customer insertCustomer(Customer customer) {
        return customerRepository.save(customer);
    }

    @Override
    public boolean isCustomerExistsWithEmail(String email) {
        return customerRepository.existsCustomerByEmail(email);
    }

    @Override
    public boolean isCustomerExistsById(Long id) {
        return customerRepository.existsCustomerById(id);
    }

    @Override
    public void deleteCustomer(Long id) {
        customerRepository.deleteById(id);
    }

    @Override
    public Customer updateCustomer(Customer customer) {
        return customerRepository.save(customer);
    }

    @Override
    public Optional<Customer> findCustomerByEmail(String email) {
        return customerRepository.findCustomerByEmail(email);
    }
}
